requirements.txt
asgiref==3.4.1
Django==3.2.9
django-rest-framework==0.1.0
djangorestframework==3.12.4
Faker==9.8.1
python-dateutil==2.8.2
python-slugify==5.0.2
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
